+++ backend-rust/src/migrations/m0003-index-accounts-with-delegated-stake-to-baker-id.sql
@@ -1,0 +1,2 @@
+-- Important for quickly calculating the delegated stake to a baker pool.
+CREATE INDEX delegated_target_baker_id_index ON accounts(delegated_target_baker_id);